Improvements to overall architecture

+++ include/yadi/parser.h
@@ -36,7 +36,7 @@ enum class DataType : uint8_t
     TIME = 27,
 };
 
-void write_size(std::vector<uint8_t> &buffer, unsigned size);
+void write_size(std::vector<uint8_t> &buffer, size_t size);
 auto from_string(std::string const& str, DataType tag = DataType::STRING) -> std::vector<uint8_t>;
 auto from_bytes(std::vector<uint8_t> const& data, DataType tag = DataType::OCTET_STRING) -> std::vector<uint8_t>;
 auto to_string(std::vector<uint8_t> const& buffer) -> std::string;
